requirements.txt
Django==2.2.6
django-bootstrap-form==3.4
libmagic==1.0
Pillow==5.4.1
python-magic==0.4.15
python-magic-bin==0.4.14
pytz==2019.3
sqlparse==0.3.0
